Added a renting form
You now need to fill out a form when renting a car through eMove. The form isn't finished yet, I hope to display it using an iframe directly on the eMove page.

Need to finish adding all the fields required (a validation button too)
Need to finish the CSS of the form